update: refactored function tokenization and fixed makefile reading input of minishell
