+++ .gitignore
@@ -6,3 +6,4 @@ a.out
 minishell
 *.json
 includesn
+.nfs
