+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: brunogue <brunogue@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/22 14:55:53 by brunogue          #+#    #+#             */
-/*   Updated: 2025/05/30 17:29:09 by brunogue         ###   ########.fr       */
+/*   Updated: 2025/06/04 16:10:45 by brunogue         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -62,12 +62,19 @@ typedef struct s_cmd
 	struct s_cmd	*next;
 }	t_cmd;
 
-// FUNCTIONS
-int				check_quotes(char *input);
+//token.c
 t_token_type	find_token_type(char *str);
-t_token			*tokenization(char *input);
+t_token *tokenization(t_token *token, char *input);
+int     handle_quotes(char *input, int *i, t_token **token, t_token **current);
+void    append_token(t_token **token, t_token **current, char *value);
 void			ft_print_token(t_token *list);
-void			free_token_list(t_token *head);
+
+//parser.c
+int				check_quotes(char *input);
 int				valid_pipe(t_token *list);
+
+//free.c
+void			free_token_list(t_token *head);
+
 
 #endif
