+++ list.txt
@@ -37,3 +37,6 @@ leaks do token corrigidos pela função free_token_list;
 
 30/5
 brunogue - arrumado a tokenization, agora esta lendo caracter por caracter. Só falta refatorar ela, mas está mto tarde kk
+
+4/6
+brunogue - feito o refatoramento da tokenization e arrumado o makefile para que ele ignore a saida do minishell.
