+++ srcs/parser.c
@@ -1,17 +1,38 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   parse.c                                            :+:      :+:    :+:   */
+/*   parser.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: pvitor-l <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: brunogue <brunogue@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/28 16:32:53 by pvitor-l          #+#    #+#             */
-/*   Updated: 2025/05/28 16:58:41 by pvitor-l         ###   ########.fr       */
+/*   Updated: 2025/06/02 15:33:35 by brunogue         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
+int	check_quotes(char *input)
+{
+	int	count_quote;
+	int	count_double_quote;
+	int	i;
+
+	i = 0;
+	count_quote = 0;
+	count_double_quote = 0;
+	while (input[i] != '\0')
+	{
+		if (input[i] == QUOTE)
+			count_quote++;
+		if (input[i] == DOUBLE_QUOTE)
+			count_double_quote++;
+		i++;
+	}
+	if (!(count_quote % 2 == 0) || !(count_double_quote % 2 == 0))
+		return (0);
+	return (1);
+}
 
 int valid_pipe(t_token *list)
 {
